refactor: add controller mixin and history util
