+++ lib/screens/v2/training_pack_template_list_screen.dart
@@ -26,6 +26,7 @@ import '../../services/training_pack_template_ui_service.dart';
 import '../../models/saved_hand.dart';
 import '../../models/action_entry.dart';
 import '../../services/generated_pack_history_service.dart';
+import '../../utils/generated_pack_history_utils.dart';
 import 'package:intl/intl.dart';
 import 'package:collection/collection.dart';
 import 'package:provider/provider.dart';
@@ -60,3 +61,4 @@ import 'package:timeago/timeago.dart' as timeago;
 part 'training_pack_template_list_core.dart';
 part 'training_pack_template_io.dart';
 part 'training_pack_template_filter_panel.dart';
+part 'training_pack_template_sort_filter_controller.dart';
